requirements.txt
absl-py==1.0.0
anyio==3.6.1
appnope==0.1.2
asgiref==3.5.2
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
boto3==1.23.10
botocore==1.26.10
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.3
entrypoints==0.3
fastapi==0.78.0
filelock==3.3.2
Flask==2.0.2
flatbuffers==1.12
fonttools==4.28.5
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.6.0
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
jmespath==1.0.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.21.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.1
pygame==2.1.2
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2021.3
pyzmq==22.3.0
regex==2021.11.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.0.2
wget==3.2
wrapt==1.14.1
zipp==3.8.0
